feat: Update Git commit message guidelines and enhance PR template
- Improved documentation across various files for better clarity and usability.
