+++ README.md
@@ -24,24 +24,33 @@ This project is structured as a monorepo with:
 - **Shared libraries**: Common components, services, and models used across multiple challenges
 
 ```
-ng-coding-challenges/
-├── projects/                 # Individual challenge applications
-│   └── coding-challenges/    # Main application (challenge browser)
-│       └── src/
-│           └── app/
-│               └── challenges/
-│                   ├── challenge-01/   # Individual challenge folder
-│                   ├── challenge-02/   # Individual challenge folder
-│                   └── ...
-├── libs/                     # Shared libraries
-│   └── shared/
-│       ├── ui/               # Shared UI components
-│       ├── models/           # Shared data models
-│       └── services/         # Shared services
-└── docs/                     # Documentation
-    ├── CREATE_CHALLENGE.md   # Template for creating new challenges
-    └── CONTRIBUTION.md       # Guide for contribution
 
+ng-coding-challenges/
+├── projects/
+│   ├── ngc-shell/           # 🧩 Main Application Shell
+│   │                        # - Hosts the main layout, routing, and navigation
+│   │                        # - Entry point for all challenge category apps
+│   │                        # - Responsible for global UI (header, sidebar, etc.)
+│   │
+│   ├── ngc-core/            # ⚙️ Core Challenges Category
+│   │                        # - Contains core Angular challenges
+│   │                        # - Each challenge lives inside this category folder
+│   │
+│   ├── ngc-routing/         # 🧭 Routing Challenges Category
+│   │                        # - Focused on Angular Router-related challenges
+│   │
+│   ├── ngc-rxjs-api/        # 🔄 RxJS & API Challenges Category
+│   │                        # - Deals with RxJS patterns, API handling, observables
+│   │
+│   └── ...                  # Additional challenge categories can be added here
+│
+├── libs/
+│   └── shared/
+│       ├── models/          # 🧱 Shared TypeScript models & interfaces
+│       ├── services/        # 🔧 Common Angular services (e.g., API, storage, logging)
+│       └── ui/              # 🎨 Shared UI components (buttons, cards, layouts, etc.)
+│
+└── node_modules/
 ```
 
 See the Architecture Guide [ARCHITECTURE.md](docs/ARCHITECTURE.md) for more details.
