+++ docs/ANGULAR_BEST_PRACTICES.md
@@ -1,9 +1,11 @@
-## TypeScript Best Practices
+# TypeScript Best Practices
+
 - Use strict type checking
 - Prefer type inference when the type is obvious
 - Avoid the `any` type; use `unknown` when type is uncertain
 
 ## Angular Best Practices
+
 - Always use standalone components over NgModules
 - Do NOT set `standalone: true` inside the `@Component`, `@Directive` and `@Pipe` decorators
 - Use signals for state management
@@ -12,6 +14,7 @@
 - Do NOT use the `@HostBinding` and `@HostListener` decorators. Put host bindings inside the `host` object of the `@Component` or `@Directive` decorator instead
 
 ## Components
+
 - Keep components small and focused on a single responsibility
 - Use `input()` and `output()` functions instead of decorators
 - Use `computed()` for derived state
@@ -22,19 +25,21 @@
 - DO NOT use `ngStyle`, use `style` bindings instead
 - Always unsubscribe Observables using `takeUntilDestroyed()`.
 
-
 ## State Management
+
 - Use signals for local component state
 - Use `computed()` for derived state
 - Keep state transformations pure and predictable
 - Do NOT use `mutate` on signals, use `update` or `set` instead
 
 ## Templates
+
 - Keep templates simple and avoid complex logic
 - Use native control flow (`@if`, `@for`, `@switch`) instead of `*ngIf`, `*ngFor`, `*ngSwitch`
 - Use the async pipe to handle observables
 
 ## Services
+
 - Design services around a single responsibility
 - Use the `providedIn: 'root'` option for singleton services
 - Use the `inject()` function instead of constructor injection
