--- docs/CH-12-REQUIREMENT-GUIDE.md
@@ -1,3 +1,0 @@
-# Challenge 12 - Requirement Guide
-
-Describe the requirements, constraints, and expected inputs/outputs for **test**.
