--- docs/CH-12-SOLUTION-GUIDE.md
@@ -1,3 +1,0 @@
-# Challenge 12 - Solution Guide
-
-Document the step-by-step solution approach, reasoning, and optimizations for **test**.
