+++ docs/CODE_OF_CONDUCT.md
@@ -1,61 +1,80 @@
-# Code of Conduct
+# 🤝 Code of Conduct
 
-## Our Pledge
+## 💡 Our Pledge
 
 We as members, contributors, and leaders pledge to make participation in our
-community a harassment-free experience for everyone, regardless of age, body
+community a **harassment-free experience for everyone**, regardless of age, body
 size, visible or invisible disability, ethnicity, sex characteristics, gender
 identity and expression, level of experience, education, socio-economic status,
 nationality, personal appearance, race, religion, or sexual identity and
 orientation.
 
-We pledge to act and interact in ways that contribute to an open, welcoming,
-diverse, inclusive, and healthy community.
+We pledge to act and interact in ways that contribute to an **open, welcoming,
+diverse, inclusive, and healthy community**.
 
-## Our Standards
+---
 
-Examples of behavior that contributes to a positive environment:
+## 🌟 Our Standards
 
-* Using welcoming and inclusive language
-* Being respectful of differing viewpoints and experiences
-* Gracefully accepting constructive criticism
-* Focusing on what is best for the community
-* Showing empathy towards other community members
+### Examples of behavior that contributes to a positive environment include
 
-Examples of unacceptable behavior:
+- Using welcoming and inclusive language  
+- Being respectful of differing viewpoints and experiences  
+- Gracefully accepting constructive feedback  
+- Focusing on what is best for the community  
+- Showing empathy and kindness toward others  
 
-* The use of sexualized language or imagery and unwelcome sexual attention or advances
-* Trolling, insulting/derogatory comments, and personal or political attacks
-* Public or private harassment
-* Publishing others' private information without explicit permission
-* Other conduct which could reasonably be considered inappropriate in a professional setting
+### Examples of unacceptable behavior include
 
-## Enforcement Responsibilities
+- The use of sexualized language or imagery and unwelcome sexual attention or advances  
+- Trolling, insulting, or derogatory comments, and personal or political attacks  
+- Public or private harassment  
+- Publishing others' private information (such as physical or email address) without explicit permission  
+- Other conduct that could reasonably be considered inappropriate in a professional setting  
 
-Project maintainers are responsible for clarifying and enforcing our standards of
+---
+
+## 🧭 Enforcement Responsibilities
+
+Project maintainers are responsible for **clarifying and enforcing** our standards of
 acceptable behavior and will take appropriate and fair corrective action in
-response to any behavior that they deem inappropriate, threatening, offensive,
+response to any behavior deemed inappropriate, threatening, offensive,
 or harmful.
 
-## Scope
+---
 
-This Code of Conduct applies within all community spaces, and also applies when
-an individual is officially representing the community in public spaces.
-Examples of representing our community include using an official project email
-address, posting via an official social media account, or acting as an appointed
-representative at an online or offline event.
+## 🌐 Scope
 
-## Enforcement
+This Code of Conduct applies to **all community spaces**, including GitHub discussions,
+code reviews, issues, pull requests, and project-related events — online or offline.
+
+It also applies when an individual is **officially representing** the project in public
+spaces (e.g., using an official project email address or social media handle).
+
+---
+
+## 🚨 Enforcement
 
 Instances of abusive, harassing, or otherwise unacceptable behavior may be
-reported to the project team. All complaints will be reviewed and investigated
-promptly and fairly.
+reported to the project team via:
 
-All community leaders are obligated to respect the privacy and security of the
-reporter of any incident.
+📧 **Contact:** <manishboge@gmail.com>
 
-## Attribution
+All complaints will be reviewed and investigated promptly and fairly.  
+All community leaders are obligated to **respect the privacy and security** of the reporter.
 
-This Code of Conduct is adapted from the [Contributor Covenant](https://www.contributor-covenant.org),
-version 2.0, available at
-https://www.contributor-covenant.org/version/2/0/code_of_conduct.html.
+---
+
+## 🧾 Attribution
+
+This Code of Conduct is adapted from the  
+[**Contributor Covenant**](https://www.contributor-covenant.org), version 2.0,  
+available at  
+<https://www.contributor-covenant.org/version/2/0/code_of_conduct.html>.
+
+For answers to common questions, see the FAQ at  
+<https://www.contributor-covenant.org/faq>.
+
+---
+
+> ❤️ **Remember:** A respectful and inclusive environment allows everyone to learn, share, and grow together.
