+++ docs/GIT_COMMIT_MESSAGES.md
@@ -1,18 +1,18 @@
-### Git Commit Messages
+# Git Commit Messages
 
 * Use the present tense ("Add feature" not "Added feature")
 * Use the imperative mood ("Move cursor to..." not "Moves cursor to...")
 * Limit the first line to 72 characters or less
 * Reference issues and pull requests liberally after the first line
 * Consider starting the commit message with an applicable emoji:
-    * 🎨 `:art:` when improving the format/structure of the code
-    * 🐎 `:racehorse:` when improving performance
-    * 🚱 `:non-potable_water:` when plugging memory leaks
-    * 📝 `:memo:` when writing docs
-    * 🐛 `:bug:` when fixing a bug
-    * 🔥 `:fire:` when removing code or files
-    * 💚 `:green_heart:` when fixing the CI build
-    * ✅ `:white_check_mark:` when adding tests
-    * 🔒 `:lock:` when dealing with security
-    * ⬆️ `:arrow_up:` when upgrading dependencies
-    * ⬇️ `:arrow_down:` when downgrading dependencies
+  * 🎨 `:art:` when improving the format/structure of the code
+  * 🐎 `:racehorse:` when improving performance
+  * 🚱 `:non-potable_water:` when plugging memory leaks
+  * 📝 `:memo:` when writing docs
+  * 🐛 `:bug:` when fixing a bug
+  * 🔥 `:fire:` when removing code or files
+  * 💚 `:green_heart:` when fixing the CI build
+  * ✅ `:white_check_mark:` when adding tests
+  * 🔒 `:lock:` when dealing with security
+  * ⬆️ `:arrow_up:` when upgrading dependencies
+  * ⬇️ `:arrow_down:` when downgrading dependencies
